CASMPET-3809: Configure istio monitoring apps
Users couldn't access Istio's Kiali because it wasn't configured
to accept the Keycloak gatekeeper access control. The configuration
was lost when we switched to the new method of configuring the
IstioOperator for istio 1.6.

After some investigation I found that the old configuration
settings were done by adding them to the `values` section.

The `enabled` and `k8s` options remain under `addOnComponents`,
so I split up the prometheus configs into 2 sections.