CASMPET-3926: Stop deploying kiali add-on using istio-operator
The istio project is not going to support deploying the add-ons
through istio-operator in a future release. This change stops
configuring istio operator to deploy Kiali.

Kiali will deployed using its own operator.

A new post-upgrade hook is added that waits for the operator to get
around to deleting the kiali objects that it created.