CASMPET-4026: Workaround memory leak in tcp-stats-filter-1.6
The SLS pods are seeing a memory leak in the istio-proxy sidecar
on larger systems. There's already a bug and a fix in newer
versions of istio[0]. The workaround mentioned in the bug is to
remove the tcp-stats-filter-1.6 envoyfilter.

The istio operator will create this if it's restarted, so the
proposed fix here is to run a cronjob that just attempts to
delete the envoyfilter every 10 minutes. The memory leak doesn't
affect any services for several hours so this should be fine.