CASMPET-4179: Update chart versions to prep for release.
These charts are going to be released for csm-1.1, so update the
version to keep track of that.

The release will include

* CASMPET-3767 : Default to distroless images

We haven't done a release for csm-1.1 yet.

There was a csm-1.0 release with versions:

- cray-istio: 1.25.1
- cray-istio-deploy: 1.19.0
- cray-istio-operator: 1.19.0

(cherry picked from commit e9d72b004e255c78d5253d2b70ae43eb61848d4e)