CASMPET-5104: Updates for Istio 1.8.6
The charts are updated for Istio 1.8.6. Here's a summary of the
changes.

* cray-istio-operator

I updated the charts/istio-operator subchart from the 1.8.6
distribution.

I removed the workaround to recreate the IstioOperator CRD. This is no
longer necessary.

* cray-istio-deploy

Changed the default image tags.

* cray-istio

Changed the default image tag.

--

In addition, when upgrading from 1.7 and also disabling the add-ons
the add-ons don't get cleaned up by the operator. A hook is added
to clean up the add-on resources. There's no need to wait for the
kiali resources to be cleaned up by the operator now.