CASMPET-5231: Switch to 1.8.6-cray2 images
The default images are switched to the distroless version of
1.8.6-cray2 from 1.8.6-cray1 to pick up security fixes.

The images annotation in the Chart.yamls are also updated so that
we don't have to specify the images to provide in the manifest.

The latest base image available has a vulnerability that snyk
detects as https://snyk.io/vuln/SNYK-DEBIAN11-GLIBC-1296898 .
There is no fix available for this according to Snyk. There's no
way to ignore a single vulnerability so the only way to fix the
vulnerabilities that are in this image is to disable the snyk
scan for now.